+++ src/microbe_stage/species.h
@@ -15,8 +15,8 @@ class Species : public RegistryType {
 public:
 	double spawnDensity = 0.0;
 	Ogre::ColourValue colour;
-	std::map<unsigned int, unsigned int> startingCompounds;
-	std::map<int, unsigned int> organelles; // TODO: get a position as the key.
+	std::map<size_t, unsigned int> startingCompounds;
+	std::map<int, size_t> organelles; // TODO: get a position as the key.
 
 	Species();
